refactor : renaming files to new format
