+++ examples/services_examples/get_chain_tip_header/get_chain_tip_header.go
