+++ examples/services_examples/get_current_block_height/get_current_block_height.go
