+++ examples/services_examples/get_merkle_path_for_tx/get_merkle_path_for_tx.go
