+++ examples/services_examples/get_rawtx_from_txid/get_rawtx_from_txid.md
