+++ examples/wallet_examples/show_address_for_tx_from_faucet/show_address_for_tx_from_faucet.md
